change: remove event_id participant uniqueness
Why: this gets unruly very quickly if someone clicks the back button; 
and, when someone goes to register next year. We don't want to recreate 
events & races YoY. We just want to add participant records. 

We will introduce a start list feature, and, implement a warning box on 
step 2 of the form if we already have a registration for this person for 
this race this _year_.