feature: resumeable registrations
This feature makes the registration process resumeable. It doesn't 
change the fact that once a registration is started, it is saved to the 
database whether it is completed (paid) or not. It simply reduces the 
amount of duplication that would otherwise be created by people clicking 
their browser's back button right after submitting the first step of the 
form. It also makes it easier for 'pick up where you left off' marketing 
efforts.

In essence, any time the new registration form is submitted, we first 
look to see if someone has already registered with that email address. 
If so, we take them to the last step _shown_. This could be any step, 
*including* the Confirmation step.