Expand admin dashboard charts
After the logic was updated for the Participation table to be net of
cancellations it was decided that it would make sense to update the
charts to be net of cancellations and show refunds for consistency.
While working on this Business thought the broader time horizon
offered more meaningful insight of KPIs.

- Update Participants chart logic to be net of cancellations
- Add refunds to Payments chart
- Format Payments chart as currency
- Widen the scope of charts to 9 months for broader view of KPIs