refactor app layout for fixed header (#199)
Refactor the spacer to render a section in the app layout instead of dynamically adding a class to %main. This moves the main section to start at the bottom of the header section. There was still some crowded elements after adding the covid banner spacer, most noticeably on the landing pages and events scrollspy section. Gives a more pleasant user experience.

* rename spacer class
* add fixed header section in /navigation
* render fixed header spacer in app layout
* remove margin-top from landing sections
* adjust sticky-top-offset for events overview
* adjust parallaxy class for image placement